downloader/requirements.txt
redis==5.0.8
yt-dlp==2024.8.6
python-dotenv==1.0.1